bientot la perfection

+++ decoder.h
@@ -35,6 +35,6 @@ typedef struct {
 int8_t decodeOpcode(int8_t byte);
 int16_t hex_to_dec(char *hex);
 void test_all_instructions();
-int readFile(char* path, int8_t write, FILE* output);
+int16_t readFile(char* path, int8_t write, FILE* output);
 int main(int argc, char* argv[]);
 instruction_informations* get_instruction_informations();
